+++ classes/prefix_WPgutenberg/blocks/posts/index.php
@@ -184,10 +184,7 @@ function WPgutenberg_posts_PostBuilder(array $attr, $id, $currentId){
     endif;
       // add post thumbnail
       if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_img', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-        if(function_exists('has_post_video') && has_post_video($id)):
-        else:
-          $output .= $linkOpen;
-        endif;
+        $output .= $linkOpen;
       endif;
         if(array_key_exists('postThumb', $attr) && $attr['postThumb'] !== false):
           if($attr['postType'] == 'attachment'):
@@ -203,48 +200,34 @@ function WPgutenberg_posts_PostBuilder(array $attr, $id, $currentId){
             endif;
             $output .= '<figure>' . $file . '</figure>';
           else:
-            $output .= get_the_post_thumbnail($id) ? '<figure>' . get_the_post_thumbnail($id, 'full', array("data-id" => $id)) . '</figure>' : '';
+            $output .= get_the_post_thumbnail($id) ? '<figure>' . get_the_post_thumbnail($id, 'full', array("data-id" => $id, 'callingFrom' => 'postsblock')) . '</figure>' : '';
           endif;
         endif;
       if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_img', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-        if(function_exists('has_post_video') && has_post_video($id)):
-        else:
-          $output .= $linkClose;
-        endif;
+        $output .= $linkClose;
       endif;
       // add content
       $output .= '<div class="post-content">';
         if(array_key_exists('postTextOne', $attr) && $attr['postTextOne'] !== ''):
           $output .= $attr['postTextOne'] == 'title' ? '<h4>' : '<div class="' . $attr['postTextOne'] . '">';
             if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row1', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-              if(function_exists('has_post_video') && has_post_video($id) && ($attr['postTextOne'] == 'templateMedia' || $attr['postTextOne'] == 'template')):
-              else:
-                $output .= $linkOpen;
-              endif;
+
+              $output .= $linkOpen;
             endif;
               $output .= WPgutenberg_posts_ContentRow($attr['postTextOne'], $id);
             if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row1', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-              if(function_exists('has_post_video') && has_post_video($id) && ($attr['postTextOne'] == 'templateMedia' || $attr['postTextOne'] == 'template')):
-              else:
-                $output .= $linkClose;
-              endif;
+              $output .= $linkClose;
             endif;
           $output .= $attr['postTextOne'] == 'title' ? '</h4>' : '</div>';
         endif;
         if(array_key_exists('postTextTwo', $attr) && $attr['postTextTwo'] !== ''):
           $output .= $attr['postTextTwo'] == 'title' ? '<h4>' : '<div class="' . $attr['postTextTwo'] . '">';
             if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row2', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-              if(function_exists('has_post_video') && has_post_video($id) && ($attr['postTextTwo'] == 'templateMedia' || $attr['postTextTwo'] == 'template')):
-              else:
-                $output .= $linkOpen;
-              endif;
+              $output .= $linkOpen;
             endif;
               $output .= WPgutenberg_posts_ContentRow($attr['postTextTwo'], $id);
             if(array_key_exists('postTaxonomyFilterOptions', $attr) && in_array('link_row2', $attr['postTaxonomyFilterOptions']) && !in_array('link_box', $attr['postTaxonomyFilterOptions'])):
-              if(function_exists('has_post_video') && has_post_video($id) && ($attr['postTextTwo'] == 'templateMedia' || $attr['postTextTwo'] == 'template')):
-              else:
-                $output .= $linkClose;
-              endif;
+              $output .= $linkClose;
             endif;
           $output .= $attr['postTextTwo'] == 'title' ? '</h4>' : '</div>';
         endif;