+++ template_parts/post_default.php
@@ -3,20 +3,22 @@
  * Blog template for defaults
  *
  * @author      David Voglgsang
- * @version     1.1.1
+ * @version     1.2.1
  *
 */
+if($args['callingFrom'] && $args['callingFrom'] !== ''):
+  $callingFrom = $args['callingFrom'];
+else:
+  $callingFrom = '';
+endif;
 if($args['id'] && $args['id'] !== 0):
   global $post;
   $post = get_post($args['id']);
 endif;
 
 
-if(function_exists('has_post_video') && has_post_video()):
-  $thumb = get_the_post_video();
-else:
-  $thumb = get_the_post_thumbnail();
-endif;
+
+$thumb = get_the_post_thumbnail(get_the_id(), 'full', array('callingFrom' => $callingFrom));
 
 $css = 'temp-' . get_post_type() . '-default';
 // disable content loading